test: consolidate assertions into adjacent test files
